fix: db and docker

+++ internal/controllers/threads/create.go
@@ -5,7 +5,7 @@ import (
 	"gochan/pkg/auth"
 )
 
-func CreateThread(h *repository.Repo, head, text, password, img string) (int64, error) {
+func CreateThread(h *repository.Repo, head, text, password, img string) (int, error) {
 	passwordHash := auth.CreateSum(password)
 	id, err := h.Thread().CreateThread(head, text, passwordHash, img)
 	return id, err
